'average_product_sold per cart' information is now available to the client for every product. Have implemented the code for so, in the serializers.py

+++ shoping_api_app/models.py
@@ -41,7 +41,7 @@ class Product(models.Model):
         else it's regular price is returned.
         '''
         if self.is_on_sale():
-            discounted_price = self.price * (1 - self.DISCOUNT_RATE)
+            discounted_price = float(self.price * (1 - self.DISCOUNT_RATE))
             return round(discounted_price, 2)
         return self.get_rounded_price()
